+++ tilelang/dtypes.py
@@ -1,0 +1,3 @@
+# Re-export from language.dtypes for convenient access via `from tilelang.dtypes import ...`
+from tilelang.language.dtypes import *  # noqa: F401, F403
+from tilelang.language.dtypes import dtype, AnyDType, get_tvm_dtype  # noqa: F401
