+++ tilelang/language/eager/__init__.py
@@ -1,2 +1,2 @@
 from .builder import prim_func, macro, PrimFunc, JITFunc, Ref, const  # noqa: F401
-from .dtypes import *
+from ..dtypes import *
